renaming to JsonConnectorDef and implementing basic checkConnection
