+++ connectors/cnext/_defaults/index.ts
@@ -1,9 +1,9 @@
 import type {ConnectorSchemas, ConnectorServer} from '../../../kits/cdk'
-import type {ConnectorDef} from '../def'
+import type {JsonConnectorDef} from '../def'
 import {generateOAuth2Server} from './oauth2'
 
-export function generateConnectorServerV1<T extends ConnectorSchemas>(
-  connectorDef: ConnectorDef,
+export function generateConnectorServer<T extends ConnectorSchemas>(
+  connectorDef: JsonConnectorDef,
 ): ConnectorServer<T> {
   switch (connectorDef.auth_type) {
     case 'OAUTH2':
