+++ connectors/cnext/index.ts
@@ -1,10 +1,10 @@
 import {
   def as connectorGoogledrive_def,
   server as connectorGoogledrive_server,
-} from './googledrive'
+} from './googledrive/def'
 
 // generated by _generateConnectorLists.ts. Do not modify by hand
-export type {ConnectorDef} from './def'
+export type {JsonConnectorDef as ConnectorDef} from './def'
 export type {AuthType} from './def'
 
 export {connectorGoogledrive_server, connectorGoogledrive_def}
